+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Methods in the `Base` effect now accept `Base` layers instead of `Visual` layers.
+
 ### Fixed
 - Layers no longer trigger infinite loops when their active states change ([#127](https://github.com/etro-js/etro/issues/127)).
 - Add missing `VisualSource` options to `Image` layer ([#128](https://github.com/etro-js/etro/pull/128)).