+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.9.1] - 2022-09-18
 ### Fixed
 - Update color types from `string` to `Color` ([#135](https://github.com/etro-js/etro/pull/135)).
 - `Image` and `Video` classes now include missing properties.
@@ -233,6 +233,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Gaussian blur
   - Transform
 
+[0.9.1]: https://github.com/etro-js/etro/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/etro-js/etro/compare/v0.8.5...v0.9.0
 [0.8.5]: https://github.com/etro-js/etro/compare/v0.8.4...v0.8.5
 [0.8.4]: https://github.com/etro-js/etro/compare/v0.8.3...v0.8.4